Clean up associations - no longer use `instance`
Previously, we were allowing factories to set up recursive associations.
Eg. User has many posts and Post has many users. We were enabling this
by passing in an `instance` property that factories could use to
reference themselves for passing into relationships. This was
problematic, though, since the instance object wasn't actually an
instance of the specified class. It was just a dummy object that we were
instantiating and then later filling in (after the association was
created). The main problems with this were:

* This wouldn't work with object types that were not object literals.
The dummy object was not an instance of the class
* The dummy object wasn't a real object of the specified type, so if
factories tried to use this object, things would break

To fix this, I removed the `instance` param, and if factories want to
set up associations, they should ensure they are not recursive. This
means it is now a little tricker to set up a belongs-to <-> belongs-to,
but is that even a thing, so much?