Allow using factories without registering
Previously, we were throwing an error when trying to use a factory without registering it
even if the factory didn't access the 'factories' object which gets injected when registering.
This updates so that the error is only thrown if 'factories' is used in the unregistered factory.
This makes it easier to define one-off factories in individual tests, since they no longer need to be registered