Preserve the object passed to build function
Previously when an object with an explicit property of undefined was
passed, it was being ignored and substituted with the object property from the
factory definition. This was not the correct behaviour because the passed property of user was lost.

Now this is preserved and does not get automatically substituted.