Update FactoryInstance with new generics
Previously, we were tracking the type when using traits/extending/params by
updating the type of T. This was problematic since the factory was initially
defined to work with the original type, T, and not the updated type. Now, we
have a new generic with the 'ExtensionParams', which are updated whenever a
factory is extended or a trais is used or factory.params is called. We were
already storing those params and passing them along when creating new factories
but were not typing them separately.

Also added a new generic, BuildType, which allows us to use custom naming when calling
extend. Otherwise, would not have a way to track that name