Fix merging of `Buffer` objects
Previously, if you attempted to `build`/`create` an object whose properties contained `Buffer`s, they would be converted from the native `Buffer` type into an array (and thus creating an _incorrect_ value). This is caused by the use of lodash's `merge` implementation (see https://github.com/lodash/lodash/issues/2964).

The recommended solution from the lodash project is to define a merge `customizer` to specifically handle this situation, which is what we do here.